Optimise bit banging for speed
The existing version was implemented just as a first attempt, to see if
I could get the whole thing to work. On an ESP32, the existing version
was able to hit a clock frequency of 1.7Mhz. The version included in
this commit now reaches 2.941Mhz.

The refresh frequency of the display can be calculated by applying the
following formula:

f = clock frequency / (columns * addressable rows * colour depth)

There are 16 addressable rows on my panels, and I have between 64 and
256 columns, depending on how many panels I decide to daisy-chain. With
a single panel, and 1 bit per channel (on or off, so 9 colours), this
gives me a maximum refresh frequency of 2.8Khz. If I use 4 panels, and
want to have 8 bits per channel (255 values, so 16M colours), that's
just under 3Hz.

One obvious way forward is to use Binary Code Modulation[1], which
reduces the number of iterations required from 255 to 8. Assuming I'm
able to perfectly map things, that would still only give me just shy of
90hz of refresh rate for the display. I was hoping to stay above 240Hz
at a minimum.

[1]: http://www.batsocks.co.uk/readme/art_bcm_3.htm